+++ packages/langchain/src/orchestration/client.test.ts
@@ -1,0 +1,106 @@
+import { constructCompletionPostRequest } from '@sap-ai-sdk/orchestration/internal.js';
+import { jest } from '@jest/globals';
+import nock from 'nock';
+import {
+  mockClientCredentialsGrantCall,
+  mockDeploymentsList,
+  mockInference,
+  parseMockResponse
+} from '../../../../test-util/mock-http.js';
+import { OrchestrationClient } from './client.js';
+import type {
+  CompletionPostResponse,
+  OrchestrationModuleConfig
+} from '@sap-ai-sdk/orchestration';
+
+jest.setTimeout(30000);
+
+describe('orchestration service client', () => {
+  let mockResponse: CompletionPostResponse;
+  beforeEach(async () => {
+    mockClientCredentialsGrantCall();
+    mockDeploymentsList({ scenarioId: 'orchestration' }, { id: '1234' });
+    mockResponse = await parseMockResponse<CompletionPostResponse>(
+      'orchestration',
+      'orchestration-chat-completion-success-response.json'
+    );
+  });
+
+  afterEach(() => {
+    nock.cleanAll();
+  });
+
+  function mockInferenceWithResilience(resilience: {
+    retry?: number;
+    delay?: number;
+  }) {
+    mockInference(
+      {
+        data: constructCompletionPostRequest(config, { messagesHistory: [] })
+      },
+      {
+        data: mockResponse,
+        status: 200
+      },
+      {
+        url: 'inference/deployments/1234/completion'
+      },
+      resilience
+    );
+  }
+
+  const config: OrchestrationModuleConfig = {
+    llm: {
+      model_name: 'gpt-35-turbo-16k',
+      model_params: { max_tokens: 50, temperature: 0.1 }
+    },
+    templating: {
+      template: [{ role: 'user', content: 'Hello!' }]
+    }
+  };
+
+  it('returns successful response when maxRetries equals retry configuration', async () => {
+    mockInferenceWithResilience({ retry: 2 });
+
+    const client = new OrchestrationClient(config, {
+      maxRetries: 2
+    });
+
+    expect(await client.invoke([])).toMatchSnapshot();
+  });
+
+  it('throws error response when maxRetries is smaller than required retries', async () => {
+    mockInferenceWithResilience({ retry: 2 });
+
+    const client = new OrchestrationClient(config, {
+      maxRetries: 1
+    });
+
+    await expect(client.invoke([])).rejects.toThrow(
+      'Request failed with status code 500'
+    );
+  });
+
+  it('throws when delay exceeds timeout', async () => {
+    mockInferenceWithResilience({ delay: 2000 });
+
+    const client = new OrchestrationClient(config);
+
+    const response = client.invoke([], { timeout: 1000 });
+
+    await expect(response).rejects.toThrow(
+      expect.objectContaining({
+        stack: expect.stringMatching(/Timeout/)
+      })
+    );
+  });
+
+  it('returns successful response when timeout is bigger than delay', async () => {
+    mockInferenceWithResilience({ delay: 2000 });
+
+    const client = new OrchestrationClient(config);
+
+    const response = await client.invoke([], { timeout: 3000 });
+    expect(response).toMatchSnapshot();
+  });
+});