+++ sample-code/src/server.ts
@@ -20,7 +20,8 @@ import {
   chatCompletionStreamWithJsonModuleConfig as orchestrationChatCompletionStreamWithJsonModuleConfig,
   orchestrationGroundingHelpSapCom,
   orchestrationMaskGroundingInput,
-  orchestrationPromptRegistry
+  orchestrationPromptRegistry,
+  orchestrationResponseFormat
 } from './orchestration.js';
 import {
   getDeployments,
@@ -41,6 +42,7 @@ import {
   invoke,
   invokeToolChain
 } from './langchain-azure-openai.js';
+import { invokeChain as invokeChainOrchestration } from './langchain-orchestration.js';
 import {
   createCollection,
   createDocumentsWithTimestamp,
@@ -52,7 +54,7 @@ import {
   deletePromptTemplate
 } from './prompt-registry.js';
 import type { RetievalPerFilterSearchResult } from '@sap-ai-sdk/document-grounding';
-import type { AiApiError, AiDeploymentStatus } from '@sap-ai-sdk/ai-api';
+import type { AiDeploymentStatus } from '@sap-ai-sdk/ai-api';
 import type { OrchestrationResponse } from '@sap-ai-sdk/orchestration';
 
 const app = express();
@@ -66,6 +68,15 @@ app.get(['/', '/health'], (req, res) => {
   res.send('Hello World! 🌍');
 });
 
+function sendError(res: any, error: any, send: boolean = true) {
+  console.error(error.stack);
+  if (send) {
+    res
+      .status(error.cause?.status || 500)
+      .send(error.cause?.response?.data || error.message);
+  }
+}
+
 /* AI API */
 app.get('/ai-api/deployments', async (req, res) => {
   try {
@@ -73,11 +84,7 @@ app.get('/ai-api/deployments', async (req, res) => {
       await getDeployments('default', req.query.status as AiDeploymentStatus)
     );
   } catch (error: any) {
-    console.error(error);
-    const apiError = error.response.data.error as AiApiError;
-    res
-      .status(error.response.status)
-      .send('Yikes, vibes are off apparently 😬 -> ' + apiError.message);
+    sendError(res, error);
   }
 });
 
@@ -90,11 +97,7 @@ app.get('/ai-api/deployments-with-destination', async (req, res) => {
       )
     );
   } catch (error: any) {
-    console.error(error);
-    const apiError = error.response.data.error as AiApiError;
-    res
-      .status(error.response.status)
-      .send('Yikes, vibes are off apparently 😬 -> ' + apiError.message);
+    sendError(res, error);
   }
 });
 
@@ -102,11 +105,7 @@ app.post('/ai-api/deployment/create', express.json(), async (req, res) => {
   try {
     res.send(await createDeployment(req.body.configurationId, 'default'));
   } catch (error: any) {
-    console.error(error);
-    const apiError = error.response.data.error as AiApiError;
-    res
-      .status(error.response.status)
-      .send('Yikes, vibes are off apparently 😬 -> ' + apiError.message);
+    sendError(res, error);
   }
 });
 
@@ -114,11 +113,7 @@ app.patch('/ai-api/deployment/batch-stop', express.json(), async (req, res) => {
   try {
     res.send(await stopDeployments(req.body.configurationId, 'default'));
   } catch (error: any) {
-    console.error(error);
-    const apiError = error.response.data.error as AiApiError;
-    res
-      .status(error.response.status)
-      .send('Yikes, vibes are off apparently 😬 -> ' + apiError.message);
+    sendError(res, error);
   }
 });
 
@@ -129,11 +124,7 @@ app.delete(
     try {
       res.send(await deleteDeployments('default'));
     } catch (error: any) {
-      console.error(error);
-      const apiError = error.response.data.error as AiApiError;
-      res
-        .status(error.response.status)
-        .send('Yikes, vibes are off apparently 😬 -> ' + apiError.message);
+      sendError(res, error);
     }
   }
 );
@@ -142,11 +133,7 @@ app.get('/ai-api/scenarios', async (req, res) => {
   try {
     res.send(await getScenarios('default'));
   } catch (error: any) {
-    console.error(error);
-    const apiError = error.response.data.error as AiApiError;
-    res
-      .status(error.response.status)
-      .send('Yikes, vibes are off apparently 😬 -> ' + apiError.message);
+    sendError(res, error);
   }
 });
 
@@ -154,11 +141,7 @@ app.get('/ai-api/models', async (req, res) => {
   try {
     res.send(await getModelsInScenario('foundation-models', 'default'));
   } catch (error: any) {
-    console.error(error);
-    const apiError = error.response.data.error as AiApiError;
-    res
-      .status(error.response.status)
-      .send('Yikes, vibes are off apparently 😬 -> ' + apiError.message);
+    sendError(res, error);
   }
 });
 
@@ -166,24 +149,18 @@ app.get('/ai-api/models', async (req, res) => {
 app.get('/azure-openai/chat-completion', async (req, res) => {
   try {
     const response = await chatCompletion();
-    res.send(response.getContent());
+    res.header('Content-Type', 'text/plain').send(response.getContent());
   } catch (error: any) {
-    console.error(error);
-    res
-      .status(500)
-      .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
+    sendError(res, error);
   }
 });
 
 app.get('/azure-openai/chat-completion-with-destination', async (req, res) => {
   try {
     const response = await chatCompletionWithDestination();
-    res.send(response.getContent());
+    res.header('Content-Type', 'text/plain').send(response.getContent());
   } catch (error: any) {
-    console.error(error);
-    res
-      .status(500)
-      .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
+    sendError(res, error);
   }
 });
 
@@ -226,10 +203,7 @@ app.get('/azure-openai/chat-completion-stream', async (req, res) => {
       res.write(`  - Total tokens: ${tokenUsage.total_tokens}\n`);
     }
   } catch (error: any) {
-    console.error(error);
-    res
-      .status(500)
-      .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
+    sendError(res, error, false);
   } finally {
     res.end();
   }
@@ -245,10 +219,7 @@ app.get('/azure-openai/embedding', async (req, res) => {
       res.send('Number crunching success, got a nice vector.');
     }
   } catch (error: any) {
-    console.error(error);
-    res
-      .status(500)
-      .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
+    sendError(res, error);
   }
 });
 
@@ -264,25 +235,37 @@ app.get('/orchestration/:sampleCase', async (req, res) => {
       outputFiltering: orchestrationOutputFiltering,
       requestConfig: orchestrationRequestConfig,
       fromJson: orchestrationFromJson,
-      image: orchestrationChatCompletionImage
+      image: orchestrationChatCompletionImage,
+      responseFormat: orchestrationResponseFormat,
+      maskGroundingInput: orchestrationMaskGroundingInput
     }[sampleCase] || orchestrationChatCompletion;
 
   try {
     const result = (await testCase()) as OrchestrationResponse;
     if (sampleCase === 'inputFiltering') {
-      res.send('Input filter applied successfully');
+      res
+        .header('Content-Type', 'text/plain')
+        .send(
+          `Input filter applied successfully with response:\n${JSON.stringify(result, null, 2)}`
+        );
     } else if (sampleCase === 'outputFiltering') {
-      res.send(
-        `Output filter applied successfully with threshold results: ${JSON.stringify(result.data.module_results.output_filtering!.data!)}`
-      );
+      res
+        .header('Content-Type', 'text/plain')
+        .send(
+          `Output filter applied successfully with threshold results:\n${JSON.stringify(result.data.module_results.output_filtering!.data!, null, 2)}`
+        );
+    } else if (sampleCase === 'responseFormat') {
+      res
+        .header('Content-Type', 'text/plain')
+        .send(
+          `Response format applied successfully with response:\n${JSON.stringify(result, null, 2)}`
+        );
     } else {
-      res.send(result.getContent());
+      console.log(JSON.stringify(result.data, null, 2));
+      res.header('Content-Type', 'text/plain').send(result.getContent());
     }
   } catch (error: any) {
-    console.error(error);
-    res
-      .status(500)
-      .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
+    sendError(res, error);
   }
 });
 
@@ -331,10 +314,7 @@ app.post(
         res.write(`  - Total tokens: ${tokenUsage?.total_tokens}\n`);
       }
     } catch (error: any) {
-      console.error(error);
-      res
-        .status(500)
-        .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
+      sendError(res, error, false);
     } finally {
       res.end();
     }
@@ -383,10 +363,7 @@ app.get(
         res.write(`  - Total tokens: ${tokenUsage?.total_tokens}\n`);
       }
     } catch (error: any) {
-      console.error(error);
-      res
-        .status(500)
-        .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
+      sendError(res, error, false);
     } finally {
       res.end();
     }
@@ -396,18 +373,23 @@ app.get(
 /* Langchain */
 app.get('/langchain/invoke', async (req, res) => {
   try {
-    res.send(await invoke());
+    res.header('Content-Type', 'text/plain').send(await invoke());
   } catch (error: any) {
-    console.error(error);
-    res
-      .status(500)
-      .send('Yikes, vibes are off apparently 😬 -> ' + error.request.data);
+    sendError(res, error);
   }
 });
 
 app.get('/langchain/invoke-chain', async (req, res) => {
   try {
-    res.send(await invokeChain());
+    res.header('Content-Type', 'text/plain').send(await invokeChain());
+  } catch (error: any) {
+    sendError(res, error);
+  }
+});
+
+app.get('/langchain/invoke-chain-orchestration', async (req, res) => {
+  try {
+    res.send(await invokeChainOrchestration());
   } catch (error: any) {
     console.error(error);
     res
@@ -418,23 +400,17 @@ app.get('/langchain/invoke-chain', async (req, res) => {
 
 app.get('/langchain/invoke-rag-chain', async (req, res) => {
   try {
-    res.send(await invokeRagChain());
+    res.header('Content-Type', 'text/plain').send(await invokeRagChain());
   } catch (error: any) {
-    console.error(error);
-    res
-      .status(500)
-      .send('Yikes, vibes are off apparently 😬 -> ' + error.request.data);
+    sendError(res, error);
   }
 });
 
 app.get('/langchain/invoke-tool-chain', async (req, res) => {
   try {
-    res.send(await invokeToolChain());
+    res.header('Content-Type', 'text/plain').send(await invokeToolChain());
   } catch (error: any) {
-    console.error(error);
-    res
-      .status(500)
-      .send('Yikes, vibes are off apparently 😬 -> ' + error.request.data);
+    sendError(res, error);
   }
 });
 
@@ -468,10 +444,7 @@ app.get(
 
       res.end();
     } catch (error: any) {
-      console.error(error);
-      res
-        .status(500)
-        .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
+      sendError(res, error);
     }
   }
 );
@@ -523,10 +496,7 @@ app.get('/document-grounding/invoke-retrieve-documents', async (req, res) => {
 
     res.end();
   } catch (error: any) {
-    console.error(error);
-    res
-      .status(500)
-      .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
+    sendError(res, error);
   }
 });
 
@@ -535,27 +505,11 @@ app.get(
   async (req, res) => {
     try {
       const groundingResult = await orchestrationGroundingHelpSapCom();
-      res.send(groundingResult.getContent());
-    } catch (error: any) {
-      console.error(error);
       res
-        .status(500)
-        .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
-    }
-  }
-);
-
-app.post(
-  '/document-grounding/invoke-orchestration-mask-grounding-input',
-  async (req, res) => {
-    try {
-      const groundingResult = await orchestrationMaskGroundingInput();
-      res.send(groundingResult.getContent());
+        .header('Content-Type', 'text/plain')
+        .send(groundingResult.getContent());
     } catch (error: any) {
-      console.error(error);
-      res
-        .status(500)
-        .send('Yikes, vibes are off apparently 😬 -> ' + error.message);
+      sendError(res, error);
     }
   }
 );