+++ packages/core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sap-ai-sdk/core
 
+## 1.9.0
+
+### Patch Changes
+
+- bfed500: [Fixed Issue] Add missing cause in the error object for failing HTTP requests by using `ErrorWithCause`, providing more context for debugging.
+
+  [Compatibility Note] Due to the introduction of `ErrorWithCause`, `AxiosError` is now wrapped inside the `cause` property.
+  For example, use `error.cause.response.data` instead of `error.response.data` to access the error response from the server.
+
+## 1.8.0
+
 ## 1.7.0
 
 ## 1.6.0