+++ packages/document-grounding/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ai-sdk/document-grounding
 
+## 1.9.0
+
+### Patch Changes
+
+- Updated dependencies [bfed500]
+  - @sap-ai-sdk/core@1.9.0
+
+## 1.8.0
+
+### Patch Changes
+
+- @sap-ai-sdk/core@1.8.0
+
 ## 1.7.0
 
 ### Patch Changes